+++ resources/sys_patch.py
@@ -760,9 +760,12 @@ set million colour before rebooting"""
             if self.constants.detected_os > self.constants.big_sur:
                 self.legacy_wifi = True
 
-        if self.model in ["MacBookPro5,1", "MacBookPro5,2", "MacBookPro5,3", "MacBookPro8,2", "MacBookPro8,3"]:
+        # if self.model in ["MacBookPro5,1", "MacBookPro5,2", "MacBookPro5,3", "MacBookPro8,2", "MacBookPro8,3"]:
+        if self.model in ["MacBookPro8,2", "MacBookPro8,3"]:
             # Sierra uses a legacy GMUX control method needed for dGPU switching on MacBookPro5,x
             # Same method is also used for demuxed machines
+            # Note that MacBookPro5,x machines are extremely unstable with this patch set, so disabled until investigated further
+            # Ref: https://github.com/dortania/OpenCore-Legacy-Patcher/files/7360909/KP-b10-030.txt
             if self.constants.detected_os > self.constants.high_sierra:
                 if self.model in ["MacBookPro8,2", "MacBookPro8,3"]:
                     # Ref: https://doslabelectronics.com/Demux.html