+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow for setting custom SIP values via TUI
 - Drop `CSR_ALLOW_EXECUTABLE_POLICY_OVERRIDE` requirement for root patching
   - Lowers default SIP Disabled value to 0xA03
+- Remove legacy GMUX patch set for MacBookPro5,x machines due to instabilites
 
 ## 0.3.0
 - Fix Nvidia Tesla Acceleration in Monterey Beta 7+